// LMS.Angular/src/app/models/Trainer.ts
export class Trainer {
    image: string | undefined;
    nationalSecurutiyNumber: string | undefined;
    fname: string | undefined;
    lname: string | undefined;
    phoneNumber: string | undefined;
    email: string | undefined;
    active: boolean | undefined;
    salary:number|undefined;


}


export class Employee {
  employeeId: number | undefined;
  username: string | undefined;
  fName: string | undefined;
  lName: string | undefined;
  email: string | undefined;
  employeeImage: string | undefined;
  phoneNumber: number | undefined;
  nationalSecurutiyNumber:number|undefined;
  roleName:number|undefined;

/**
 *      public int EmployeeId { get; set; }
        public string Username { get; set; }
        public string FName { get; set; }
        public string LName { get; set; }
        public string Email { get; set; }
        public string EmployeeImage { get; set; }
        public string PhoneNumber { get; set; }
        public double BasicSalary { get; set; }
        public string NationalSecurutiyNumber { get; set; }
        public string Password { get; set; }
        public int RoleId { get; set; }
        public string RoleName { get; set; }
 */
}